MODIFIED: Another cleanup session. MODIFIED: Remaining item charges are now sent as signed to the database. ( PostgreSQL doesn't have unsigned and if sent as unsigned -1 becomes ~0ul which doesn't fit into an integer. ) MODIFIED: I noticed that threads that won't shutdown when shutting the server down are in suspension, I added a resume call on Threadpool shutdown, hopefully this will remedy the problem.

+++ src/arcemu-world/Item.cpp
@@ -374,7 +374,7 @@ void Item::SaveToDB( int8 containerslot, int8 slot, bool firstsave, QueryBuffer*
     ss << ( GUID_LOPART( CreatorGUID ) ) << ",";
 
 	ss << GetStackCount() << ",";
-	ss << GetChargesLeft() << ",";
+	ss << int32( GetChargesLeft() ) << ",";
 	ss << uint32( m_uint32Values[ ITEM_FIELD_FLAGS ] ) << ",";
 	ss << random_prop << ", " << random_suffix << ", ";
 	ss << GetUInt32Value(ITEM_FIELD_ITEM_TEXT_ID) << ",";
